ENH: VolumeRenderingModuleWidget: Extend API adding one new signal
Facilitate integration of VolumeRendering in custom application by
allowing any module to synchronized its current Scalar and Display MRML
nodes to the Volume rendering ones. Once the MRML are synchronized, this
can be used to update the GUI of both modules.

Use case driving this change is SlicerAstro.

Reviewed-by: Andras Lasso <lasso@queensu.ca>
Reviewed-by: Steve Pieper <pieper@bwh.harvard.edu>

From: Davide <punzodavide@hotmail.it>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24922 3bd1e089-480b-0410-8dfb-8563597acbee