ENH: Remove test execution delay when building RUN_TESTS target manually
When RUN_TESTS target was built in VIsual Studio manually then the py_nomainwindow_MRMLCreateNodeByClassWithSetReferenceCountToOne tests crashed (as expected) and an error popup was displayed to the user. The test execution stopped until the user acknowledged the popup (and also a Windows Error Report was generated). This caused unnecessary delays in the test execution and false positive error reports.

Solved the issue by forcing CTest to disable error popups for this particular test when building RUN_TESTS manually. CTest disables error reporting for all tests when run from the command-line.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@24953 3bd1e089-480b-0410-8dfb-8563597acbee