ENH: Fixed subject hierarchy folder node names
If a subject hierarchy node is associated with a data node then their names are synchronized; and to avoid name conflict, the subject hierarchy node gets a _SubjectHierarchy postfix in its name.

If there is no associated data node then the postfix is not necessary and it's confusing when it is shown to the user (e.g., when the user is asked to select a SubjectHierarchy folder). 

This commit changes the behavior so that _SubjectHierarchy is only added if a data node is specified.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25006 3bd1e089-480b-0410-8dfb-8563597acbee