ENH: Add support for managing slice orientation presets
This commit generalizes the concept of slice orientation known as
Axial, Sagittal and Coronal so that it is possible to add, rename and
delete new ones.

Default presets (Axial, Sagittal and Coronal) are associated to a default
slice node in the application logic.

Handle orientation as 3x3 because maintaining sliceOrigin information
in matrices that are orientation presets is not relevant.

Notes:

(1) Mark Get/SetOrientationReference as protected: These methods are only
    used internally and could ultimately be removed by refactoring
    "RotateToVolumePlane()"

(2) Remove unused SetOrientationString: The method is not used in any Slicer
    extensions.

(3) Explicitly setting the orienation string to "Reformat" is
    not needed anymore because "GetOrientation()" or "GetOrientationString()"
    will return "Reformat" if the SliceToRAS matrix is not one of the preset.

(4) For sake of consistency, in the RenameSliceOrientationPreset, the
    orientation reference is also renamed.

(5) In "Copy()" method, Orientation and OrientationReference are set
    after the presets.

(6) Remove unused SetOrientationToReformat() function. At the time of this
    commit, it was used only in "SlicerRT/ExternalBeamPlanning/Logic/vtkSlicerExternalBeamPlanningModuleLogic"
    and the maintainer of the module planned on removing the call.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>
Co-authored-by: Andras Lasso <lasso@queensu.ca>

From: Davide Punzo <punzodavide@hotmail.it>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25171 3bd1e089-480b-0410-8dfb-8563597acbee