BUG: Fixed crash when saving a scene after deleting storable nodes
How to reproduce: create a storable node (without storage node), create a scene view, delete the storable node, save the scene -> Slicer crashes.
The root cause of the problem was that missing storage node for was was added to the main scene instead of the scene view.

Storage node is now added to the correct scene and also replaced a few asserts by error handling to reduce chances of crashing during scene save.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25258 3bd1e089-480b-0410-8dfb-8563597acbee