BUG: Fixed reading of bspline transforms
Bspline transforms were interpreted as inverse bspline transforms because of an incorrect
transform class type check in vtkITKTransformConverter::SetVTKBSplineParametersFromITKGeneric.
The safest way of checking of class type would be dynamic_cast, but for now only matching of GetNameOfClass() is tested.

This fixes issue (Could not create an instance of "InverseBSplineTransform_double_3_3") that occurred
when trying to use a bspline transform computed by a CLI in another CLI.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25365 3bd1e089-480b-0410-8dfb-8563597acbee