ENH: Added data probe support for segmentations
Now any displayable manager can supply information to the data probe if the GetDataProbeInfoStringForPosition virtual method is overridden.

Also poly data fill for segmentation 2D display was fixed. It was disabled due to being unstable, see https://app.assembla.com/spaces/slicerrt/tickets/809-segmentation-polydata-2d-fill-visualization-not-stable
The error was a very obscure one: it turns out the first argument of vtkCellArray::GetCell is not the index, but a special locator position, which fact should be emphasized in the documentation at least.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25379 3bd1e089-480b-0410-8dfb-8563597acbee