BUG: Slice view controller segmentation selector selects segmentation
It was quite confusing that even if there was a segmentation node in the scene, the selector in the slice view controllers showed None selection until the user changed it (note that the segmentation selector works differently than the volume selectors, as instead of setting there what to show on that layer, it can be used to change the visibility options of any segmentation, as multiple of those can be shown at the same time).
Addition of selectNodeUponCreation was not the change that fixed the issue, but still committing, because it may be useful later.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25436 3bd1e089-480b-0410-8dfb-8563597acbee