ENH: Allow display transform glyphs at markup positions
It is now possible to show where a markup point is transformed to using transform visualization.
Select "Glyph" visualization mode, open "Advanced" section and set "Source points".
Glyphs are updated in real-time as markups are moved, so this can be also used for probing the transformation
by drag-and-dropping markups in 2D and 3D views.

Transform logic needed access to Markups node, therefore module logic was moved to Modules/Loadable/Transforms/Logic.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25551 3bd1e089-480b-0410-8dfb-8563597acbee