ENH: Add further rules to DICOM patcher module
Added a rule to change DCMTK's private storage SOP class UID to CT storage,
because files with unknown (private) class UID are rejected by ITK's GDCM DICOM reader.

It is now possible to force patient name and IDs to be the same for all files in a directory
(disabled by default).


git-svn-id: http://svn.slicer.org/Slicer4/trunk@25641 3bd1e089-480b-0410-8dfb-8563597acbee