ENH: Consolidate AutoScalarRange & ScalarRangeFlag
AutoScalarRange - in vtkMRMLDisplayNode -  would define if the
ModelDisplayNode would reset its ScalarRange to map the range
of the mesh active scalar or not (manually set by the user).

ScalarRangeFlag - in vtkMRMLDisplayNode - would define if the
ModelDisplayableManager would map scalar range based on the
active scalar if UseDataScalarRange (same that AutoScalarOn),
would use the DataType range if UseDataTypeScalarRange, would
use the color node lookup table range if UseColorNodeScalarRange,
or would use the display node scalar range (aka its OWN range
defined by AutoScalarRange value) if UseDisplayNodeScalarRange.

To consolidate those two class members which show redundancies,
they are replaced by a unique ScalarRangeFlag which can take the
following values:
* UseManualScalarRange (new, equivalent to AutoScalarRangeOff)
* UseDataScalarRange (as previously, equivalent to AutoScalarRangeOn)
* UseColorNodeScalarRange (as previously)
* UseDataTypeScalarRange (as previously)
Note the removal of `UseDisplayNodeScalarRange`.

For backward compatibility, the AutoScalarRange API is conserved
and wraps Set/GetScalarRangeFlag.

The range update based on this ScalarRangeFlag is then made in
vtkMRMLModelDisplayNode::UpdateMeshPipeline instead of
vtkMRMLModelDisplayableManager::SetModelDisplayProperty, which
simplifies greatly the code in the displayable manager by only
calling `mapper->SetScalarRange(modelDisplayNode->GetScalarRange())`,
and allows that call in vtkMRMLModelSliceDisplayableManager to
now return the correct range based on that option.

vtkMRMLModelSliceDisplayableManager is also updated to instantiate
the vtkMapper in a similar way than it is made in the 3D displayable
manager for visualization consistency (deep copy of the lut, etc...)

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25663 3bd1e089-480b-0410-8dfb-8563597acbee