ENH: Made Add data dialog more keyboard-friendly
It was very inconvenient that after drag-and-dropping or selecting data for loading in "Add data" dialog, user had to click on OK button.
Hitting Enter key opened the "Add directory" window instead of loading the data.

Changed behavior so that Enter always loads the listed data sets.

Hitting space bar clicks the button that has the focus. If somebody wants to load multiple files, then after a file is selected for loading
the user has to hit space bar to add more files. Once all the files are selected the user hits Enter key.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25674 3bd1e089-480b-0410-8dfb-8563597acbee