BUG: Fix model display pipeline calls
The ThresholdFilter and AssignAttribute filters do not need to call
Update() when they are modified: VTK will take care of running the
filters when data is requested by the renderer.

Still need to call update in UpdateScalarRange() to be able to
retrieve the scalar range of the AssignAttribute filter output mesh
if UseDataScalarRange, as well as in GetOutputMesh().

Fixed tests:
* qSlicerModelsModuleWidgetTest
* qMRMLModelDisplayNodeWidgetTest1

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25679 3bd1e089-480b-0410-8dfb-8563597acbee