ENH: Moved shortcut and lightbox helper functions into segment editor widget
installKeyboardShortcuts, uninstallKeyboardShortcuts, turnOffLightboxes helper functions were previously implemented in SegmentEditor module,
but they are very commonly needed, therefore it is better to make it available in the segment editor widget where other modules can access them.

Also slightly changed the effect selector shortcuts: keys 1, 2, ...0 selects one of the first ten effects. If the selected effect is already
active then the select gets deactivated. This is useful to be able to quickly disable and re-enable an effect, for example, when there is a need
to rotate a 3D view while using the scissors effect.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25715 3bd1e089-480b-0410-8dfb-8563597acbee