ENH: Added long needed basic features made possible by new subject hierarchy design
- Re-ordering and expanded state are saved in the items, and restored on update
- Root item handling
- Subject hierarchy python scripted plugins
- Various bugfixes related to the above features and subject hierarchy core

-----------------------------------------------
Details:

BUG: Fixed subject hierarchy tree root item handling

ENH: Re-select items after drag&drop in subject hierarchy

BUG: Fixed subject hierarchy Register plugin

BUG: Fixed subject hierarchy reordering

It failed when an item was moved up not down

BUG: Fixed issues related to deleting subject hierarchy items

Also fixed the feature in DICOM plugin for converting folders into patients and studies

BUG: Fixed segmentation reparenting and visibility icon updates in SH

BUG: Fixed show/hide in subject hierarchy

ENH: Simplified creating subject hierarchy items

Instead of the Swiss army knife function that could be used to create subject hierarchy items for both data nodes and hierarchies (folder, study, etc.) there are two separate methods for CreateItem. One for creating the pure hierarchy items (folder etc.) and the other for the ones corresponding to data nodes. The latter is for internal use and should not be called.

Also fixed basic DICOM loading

ENH: Deploy scripted plugin in SubjectHierarchyPlugins folder

Previous commit deployed them in individual folders, but for consistency and simplicity, a common folder is better. It is achieved by having the __init__ script import all plugin python files dynamically

ENH: Added python scripted subject hierarchy plugins

For SegmentEditor (that replaces the core SegmentPlugin) and SegmentStatistics. Both are mostly sample plugins, to demonstrate how to write scripted plugins, and as a proof of concept for multiple scripted plugins built.
The reason why SegmentEditorSubjectHierarchyPlugins subfolder was chosen for source and install path: 1. If the plugin is added with the module file, it is attempted to be recognized as a module and will throw error 2. If a common SubjectHierarchyPlugins folder is created in each module having a scripted plugin, the __init__ files are not merged, but the last one overwrites the previous ones, and only one plugin is exposed.

ENH: Added subject hierarchy item info in SH module widget

It contains a label that displays the item's self print string.
Level and owner plugin info was removed from the tooltip, as the info box shows that, and it's not needed for normal usage (mostly for debugging)

BUG: Root item restored after complete subject hierarchy update

ENH: Subject hierarchy item expanded state handling added

BUG: Prevent crash when closing Slicer

BUG: Allow dropping subject hierarchy items on empty area

In which case it is reparented under the scene

ENH: Added item reordering support

Also fixed model update when item is reparented programmatically


git-svn-id: http://svn.slicer.org/Slicer4/trunk@25751 3bd1e089-480b-0410-8dfb-8563597acbee