ENH: Add backward compatibility support for old scenes containing SH
New node type vtkMRMLSubjectHierarchyLegacyNode is introduced, into which the subject hierarchy nodes from old MRML scenes are loaded, so that their information can be accessed and the hierarchy reconstructed. The new subject hierarchy nodes write a subject hierarchy version node attribute into the MRML scene on exporting. Most functions of the old subject hierarchy node were removed in the legacy node, and nly those methods are kept in this node from the old one that may be necessary to extract or use information loaded into this not and are not propagated into the new hierarchy (such as UIDs). Also, nested association support is also removed (because it never became completely stable), so if a scene is loaded where a data node is associated to the subject hierarchy node using nested association, then that node will not be properly represented in the hierarchy.


git-svn-id: http://svn.slicer.org/Slicer4/trunk@25760 3bd1e089-480b-0410-8dfb-8563597acbee