ENH: Improved and simplified volume and labelmap visibility from subject hierarchy
Instead of the apparently hard-to-grasp way volumes and labelmaps were shown and hidden from subject hierarchy, a new and simplified way was implemented. Now if a non-labelmap volume is shown or hidden, then it is simply shown in the background layer of all slice views, replacing any background volume shown previously. Labelmaps work the same way, except on the label layer. This, besides the simplified and more robust operation, has a few advantages: 1) It doesn't use the selection node any more, meaning that its inconsistencies (http://www.na-mic.org/Bug/view.php?id=3551) are now irrelevant from this use case, 2) It was reported that show/hide changed the position of the slice views, which is now preserved, 3) Complete synchronization of actual volume visibility and visibility icon states in subject hierarchy is now achieved

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25866 3bd1e089-480b-0410-8dfb-8563597acbee