ENH: Made easier to save DICOM tags back to subject hierarchy items
Now the save tags checkbox in DICOM export dialog is on by default, and it saves the edited tags after each edit (if the checkbox is on), not just on export. Thus if the user edits DICOM tags but cancels the dialog to do something before export, the tags are preserved and they don't have to type it in again.

Also added a check in volumes SH plugin in case Slicer is started in no-main-window mode

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25914 3bd1e089-480b-0410-8dfb-8563597acbee