BUG: Update to latest CTK to fix test errors
The new CTK version contains this fix for multi-value arguments handling in launcher.
When a multi-value argument was passed to the command-line parser, all subsequent parameters were considered part of the argument.
This lead to ignoring --python-script argument when --additional-module-paths argument was specified before.

The patch fixes these Slicer test errors:
* py_nomainwindow_SlicerTestingExitFailureTest
* py_nomainwindow_SlicerUnitTestWithErrorsTest
* py_SlicerTestingExitFailureTest).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@25970 3bd1e089-480b-0410-8dfb-8563597acbee