BUG: Remove explicit path to git executable for GitPython (Windows)
On Windows, remove the explicit path to the git executable for
GitPython. On modern versions of Git for Windows, the specified path
(/bin/git) is not correct. This results in errors when running the
command line extension wizard, for example when publishing an extension:

    $ slicerExtensionWizard --publish ~/Slicer-MyExtension/
    ...
    git.exc.GitCommandNotFound: Cmd('C:/Program Files/Git/usr/bin/git') not found due to: WindowsError('[Error 2] The system cannot find the file specified')

When the GIT_PYTHON_GIT_EXECUTABLE environment variable is not set,
GitPython uses the git executable found in the path. Since the user is
expected to run slicerExtensionWizard from Git Bash, it's reasonable to
assume that the correct git executable is in the path.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@26012 3bd1e089-480b-0410-8dfb-8563597acbee