ENH: Made auto-rotate slice to camera normal more robust
Auto-rotate slice to camera normal only works if the reformat widget is displayed.
In Reformat module, auto-rotate is shown as a checkbox on "Normal to camera" button - very far from show/hide reformat widget, so when the user just clicks that checkbox it's not obvious why the feature does not work.

Solution: turn on slice reformat widget when user clicks on the checkbox on "Normal to camera" button in Reformat module.

Another small bug was that when the user clicks the "Normal to camera" button and checkbox is checked then the button becomes unpushed, but the checkbox is still checked, so it is not clear if auto-rotate is still active or not.

Solution: changed behavior so that when the user clicks the "Normal to camera button" and the checkbox was checked, then the button becomes unpushed and the checkbox becomes unchecked, too.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@26073 3bd1e089-480b-0410-8dfb-8563597acbee