ENH: enable selection of window/level presets in Volumes
Previously these had not been set by the dicom reader so the
box was always empty and non-functional.

Now the scalar volume plugin populates the displaynode
and these values are shwon in the volumes information widget.
When the user doubleclicks on an entry the display is set
accordingly.

From: Steve Pieper <pieper@isomics.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@26113 3bd1e089-480b-0410-8dfb-8563597acbee