ENH: Make CLI output update more robust
After a CLI module is executed, nodes that have references to other nodes are updated:
their parent transform are set to the referenced transform node; and/or their subject hierarchy
parent item is set to be the same as the parent of the referenced node.

This scene updated relied on a Qt timer that performed the updates "right after" CLI execution
was completed. This mechanism was not reliable because the timer required the user to
force background event processing for a while after the CLI execution was reported to be completed,
but this workaround was not reliable (see for example this report:
https://discourse.slicer.org/t/issues-modifying-the-transform-tree/569/13).

Changed the mechanism to use the same event queue that is used for processing reading of results
back into the scene. This way it can be assured that by the time the CLI execution is reported
to be complete, all the scene updates are completed as well.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@26145 3bd1e089-480b-0410-8dfb-8563597acbee