BUG: Fixed subject hierarchy tree view and combobox filter bug
If the scene item was rejected initially due to a filter, it caused (for unknown reason) the scene item not to be examined later when new items arrived, and so the other items that fulfilled the filters were not shown either. Worth noting that if this did not happen (and the scene was not rejected), then the filters worked properly. By making sure the scene is always accepted, new data appears as expected.

Fixes #4428
https://issues.slicer.org/view.php?id=4428

git-svn-id: http://svn.slicer.org/Slicer4/trunk@26361 3bd1e089-480b-0410-8dfb-8563597acbee