ENH: Map labelmap items to segment terminology when importing
When importing a labelmap to a segmentation, the labels in the labelmap are matched to the 3dSlicerLabel attributes in the selected terminology. Terminology can be selected in the Advanced section in Import/Export section of Segmentations module. If match is made, then the found terminology is assigned to the segment, otherwise the first entry in the terminology is set. If the segment name cannot be found in the labelmap (segment name is searched in the color table to find corresponding label), then the terminology entry is left as is (which may be the default Tissue/Tissue, or what the user manually specified). This occurs if the user imports to a non-empty segmentation.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@26371 3bd1e089-480b-0410-8dfb-8563597acbee