ENH: Add option for popup vertical alignment with subjectHierarchyComboBox
The default alignment shifts the popup treeview so that the selected
item overlays with the combobox.

The option allows to turn this behavior off, to be able to get the same
behavior than with a qMRMLNodeComboBox, where the menu appears below the
combobox.

From: agirault <agirault@3bd1e089-480b-0410-8dfb-8563597acbee>

git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer-4-8@26564 3bd1e089-480b-0410-8dfb-8563597acbee