ENH: Define annotation ROI widget handle size relative to window size
The ROI widget handle was too small in HiDPI screens.
RenderWindow cannot return screen size or DPI reliably, so window size was the most robust absolute size that the handle size could be related to.
By switching to one-up slice view, the handle gets bigger, too, which may make it easier to hit it on smaller screens.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@26988 3bd1e089-480b-0410-8dfb-8563597acbee