BUG: Fix loading of Slicer style on macOS using installed version of Slicer
This commit partially reverts r26967 (ENH: Initialize application
style using its name) and explicitly instantiates and sets the style.

This is required because the Qt plugin providing the custom styles fails
to load before the application is constructed.

This is explained  because the hints allowing a successful load of the plugin
are either set in the application constructor or after it is constructed:

* qt.conf available in the application resources is read in the application constructor

* QT_PLUGIN_PATH is only set after the application is instantiated
  by using the logic reading the SlicerLauncherSettings.ini

For these reason, directly instantiating the style address the problem.

That said, to address the root cause and have the environment initialized
before constructing the application (and have QT_PLUGIN_PATH set already),
an other approach would be to:

1) Implement a method allowing to get the application file path on all platforms
   before the QApplication is constructed.

2) Move the logic reading the launcher settings before the application is initialized.

Useful references are:
* https://issues.slicer.org/view.php?id=4515
* http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=24465
* https://forum.qt.io/topic/50305/qt-5-3-and-5-4-mac-application-crashes-on-startup-after-using-macdeployqt-app-unable-to-load-plugin-cocoa/8


Reported-by: Ron Kikinis <kikinis@bwh.harvard.edu>
Reported-by: Steve Pieper <pieper@bwh.harvard.edu>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27030 3bd1e089-480b-0410-8dfb-8563597acbee