BUG: DriverScript: Fix handling of CDash project associated with drop site
This commit fixes incorrect use of the variable CTEST_PROJECT_NAME,
it was previously set under the assumption it would be used to select
the CDash project associated with the dashboard submission.

Setting CTEST_PROJECT_NAME was a no-op, it was systematically overridden
with the correct value in CTestConfig. The correct value corresponds to the
name of the top-level CMake project associated with the project being
configured, built and tested.

To address the problem, this commit introduces the variable CDASH_PROJECT_NAME
and use it to ultimately select the appropriate CDash project when
setting the variable CTEST_DROP_LOCATION.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27058 3bd1e089-480b-0410-8dfb-8563597acbee