BUG: Fix segment color in 3D when polydata has active array
When polydata had an active array (which may be generated as a side effect by some VTK filters) then 3D display of the segment was based on that array value instead of the specified segment color.
Disabled scalar visibility in the mapper to ignore these arrays and enforce coloring by the solid color specified in the actor.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27110 3bd1e089-480b-0410-8dfb-8563597acbee