ENH: Improved segment editor scissors outside fill/erase in slice mode
When positive/negative slice clip is enabled, outside fill/erase modified the segmentation on both sides of the slice plane.
Improved the behavior so that only the selected side of the slice is filled/erased, because this is generally more useful behavior.
If needed, the other side of the slice can be easily erased/filled with an additional scissor operation.


git-svn-id: http://svn.slicer.org/Slicer4/trunk@27112 3bd1e089-480b-0410-8dfb-8563597acbee