BUG: Ensure markup is visible after clicking in 3d in persistent mode
After toggling persistent mode, placing a fiducial in 3D with Slice in the
background caused the fiducial to visible right way but clicking anywhere
else (without any slice in the background) was not causing the fiducial to
appear.

This commit ensures the rendering pipeline is triggered after adding
a fiducial in persistent mode.

Fixes https://issues.slicer.org/view.php?id=4512

Tested-by: Davide Punzo <punzodavide@hotmail.it>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27122 3bd1e089-480b-0410-8dfb-8563597acbee