STYLE: slicerMacroBuildApplication: Remove explicit handling of *.nib
Starting with qt/qt@7b7c321 from 2013 [1] integrated in Qt 4.8.7, the
*.nib files are bundled as qt resources and it is not needed to copy the
file into the application package. It used to be required for no-framework
and static builds.

Interestingly the "developing-on-mac" documentation was initially
updated in 2010 [2] to mention the files were required but never got updated
following commit qt/qt@7b7c321.

[1] https://github.com/qt/qt/commit/7b7c321d5f35b6ee70db5a72d5d37e19e125d7cf
[2] https://github.com/qt/qt/commit/b2d11d1996868a4425e15bf9a14261133e999868#diff-79d96c1e781a0b3bc541fd0e3c93f0ccL162

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27124 3bd1e089-480b-0410-8dfb-8563597acbee