ENH: Reset scene name when closing the scene
After closing a scene, the URL was not reset. Therefore, when the user saved the scene, the previously loaded one could be easily overwritten.
Now clear the scene URL when the scene is closed (only when using File menu; behavior is not changed when scene.Clear() method is called directly).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27219 3bd1e089-480b-0410-8dfb-8563597acbee