COMP: Support configuring and building against custom install of Python
This commit relax requirements introduced in r26683 and allow configuring
and building Slicer against a custom installation of Python.

Note that creating a package bundling a custom installation of the
python libraries is still NOT supported, instead of failing with
an error message, an obvious warning message is now reported.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27315 3bd1e089-480b-0410-8dfb-8563597acbee