BUG: Fixed User color node reverting to File on scene reload
When a "User" color node type was saved to scene and then reloaded from scene
then its type was always set to "File". This was an issue because User type
is editable on the GUI, while File type is read-only.

Fixed by not overwriting color node type when read from file if type has been already set.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27378 3bd1e089-480b-0410-8dfb-8563597acbee