BUG: Fixed py_nomainwindow_SlicerOptionDisableSettingsTest on Windows
qWarning() did not print warning message to the console on Windows.

Fixed by adding a qSlicerCoreApplication::showConsoleMessage method for printing messages to the user on the console directly (without relying on logging macros); and using this method whenever something should be displayed to the user (instead of just written to the application log).

Logging macros should not be used for displaying messages to the user, since log messages may be harder to read (depending what log format is specified).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27433 3bd1e089-480b-0410-8dfb-8563597acbee