BUG: Show multi-select menu over eye icon as well in Data module
When right-clicking the eye icon in case of multi-selection, the "visibility" context menu was shown, but in that case the action was only applied to the first item selected. This was confusing. However not showing anything would be also confusing, so in this case the regular multi-selection menu pops up.

Re https://issues.slicer.org/view.php?id=4659

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27579 3bd1e089-480b-0410-8dfb-8563597acbee