ENH: Highlight anatomy option in terminology selector
If the anatomic region panel is collapsed but the selected type supports anatomy information then the expand button is briefly highlighted to let the user know there are extra options available. If it's not available then the expand button is disabled (if still collapsed) until a type is selected that supports it.

Override cursors have been removed, because since the change to RapidJson the lookup is immediate and they are not needed.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27589 3bd1e089-480b-0410-8dfb-8563597acbee