ENH: Make anatomic region expand button blink 3x
According to feedback the single blink of the expand button that opens the anatomic region panel in the terminology selector widget looked like a bug. Now it blinks three times, making it more obvious that it's intentional. If it's too much, then we can dial it down later, based on further feedback.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27594 3bd1e089-480b-0410-8dfb-8563597acbee