ENH: Add option to get directly referenced nodes
GetReferencedNodes and AddReferencedNodes always collected the referenced nodes recursively. An argument has been added to these functions for turning off the recursive functioning to only get the nodes that directly reference the argument node. This should not break anything because the default value for the argument is true (recursive).
Subject hierarchy tree view shows direct and recursive referenced nodes with slightly different color.

git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer-4-10@27733 3bd1e089-480b-0410-8dfb-8563597acbee