ENH: Add toggle visibility for multi-select in Data module
It is useful to be able to change the visibility of multiple selected items in the Data module (Subject hierarchy tab). So far only the Delete action was visible on multiple selection. This commit adds a "Toggle visibility" action that toggles visibility on the selected items. To avoid multiple toggling of the same item, if any of the selected item's parents is also selected, then the child item is not toggled, only the parent.

git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer-4-10@27791 3bd1e089-480b-0410-8dfb-8563597acbee