ENH: Remove subject hierarchy automatic hints
The hints that popped up automatically in the Data module's Subject hierarchy panel about using the context menu were more an annoyance than help. Automatic showing of these hints is now removed, as well as the settings entry about it. Clicking the help button still shows the hint as before.

git-svn-id: http://svn.slicer.org/Slicer4/branches/Slicer-4-10@27795 3bd1e089-480b-0410-8dfb-8563597acbee