PERF: Fix unnecessary re-adding of observers during segment editing
When a segment visibility was changed and an effect was active, layout changed notification event was sent out each time, because the logic that checked if segmentation visibility in views have changed did not work. The problem was that after saving the list of observed view node IDs, the list was cleared (probably copy-paste error).

git-svn-id: http://svn.slicer.org/Slicer4/trunk@27949 3bd1e089-480b-0410-8dfb-8563597acbee