BUG: Fixed crash when closing scene
If Models or Markup module was open when scene was closed, Slicer crashed.

The problem was that the scene in view combobox's scene model was set to nullptr (via to widget update callback) while it was being updated.

Fixed by preventing resetting the scene if it is because to the observed display node is not yet added or already removed from the scene.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28084 3bd1e089-480b-0410-8dfb-8563597acbee