ENH: Use common interactor style for slice and 3D views
Common interactor class is responsible for delegating interaction events to displayable managers (and if none of the displayable managers can process an event then emit an interaction event).

Translation from interaction event to camera manipulation events is not hardcoded anymore in the interactor style but it is moved to vtkMRMLCameraWidget.

Added support for double-click events in the interactor.

Added menu (right-click) and action (double-click) events to widgets. They are not handled yet in Slicer core.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28095 3bd1e089-480b-0410-8dfb-8563597acbee