BUG: Fix c++ tests involving layout when Slicer_BUILD_WEBENGINE_SUPPORT is OFF
This commit fixes a regression introduced in r28172 (ENH: BuildSystem: Add
Slicer_BUILD_WEBENGINE_SUPPORT option)

AA_ShareOpenGLContexts
----------------------

Enables sharing of GL context is required when the application is build
with option Slicer_BUILD_WEBENGINE_SUPPORT set to OFF and the designer
plugin "qwebengineview" is attempted to be loaded.

Without sharing of context enabled, the VTKWidget associated with
ctkVTKAbstractView emit the "resized()" signal while the widget is
adding to the layout in the constructor of ctkVTKAbstractView and this
leads to a call "forceRender()". The call of forceRender() from the constructor
ends up with segfault because the  "RequestTimer" object is not valid yet.


qMRMLLayoutManagerWithCustomFactoryTest
---------------------------------------

It also fixes qMRMLLayoutManagerWithCustomFactoryTest to use the plot node
instead of the chart node.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28182 3bd1e089-480b-0410-8dfb-8563597acbee