BUG: Fix SegmentEditorIslandsEffect to make first segment contain the largest island
Current segment selection was lost after importing the split segments, which prevented saving of largest island (the original segment remained there).
Probably this loss of selection was a side effect of some other changes.

In this fix the selection is manually restored. A more appropriate fix in the future could be modification of ImportLabelmapToSegmentationNode to preserve selection.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28472 3bd1e089-480b-0410-8dfb-8563597acbee