BUG: random image noise in screenshot captures
Due to bug in CTK, random data was returned from conversion
from vtkImageData to QImage.  This update to CTK fixes that.

In addition there's some extra cleanup to the dialog class to make the
widget more usable.

TODO: the scale factor is not working well with the latest VTK
and some users have requested a Save As button to export images
directly.

From: Steve Pieper <pieper@isomics.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28535 3bd1e089-480b-0410-8dfb-8563597acbee