BUG: avoid crash with LONG images
vtkImageInterpolator explicitly disallows reslicing of
data types that cannot be represented by a double, so
Slicer would crash when a volume of this type was loaded.

Thie change prevents the crash and generates an error message.
I'm not sure there's a better solution at this level of the
code, but it would probably make sense for the application
to generate a warning when long types of volumes are added
to the scene.

From: Steve Pieper <pieper@isomics.com>

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28563 3bd1e089-480b-0410-8dfb-8563597acbee