BUG: Fix missing labels in closed surface to binary labelmap conversion
When converting closed surface segments that have a label value > 255, the segments would not be in the labelmaps representations post-conversion.
This was because the labelmaps generated by vtkClosedSurfaceToBinaryLabelmapConversionRule are of type UNSIGNED_CHAR, and the segment label value was used for the stencil background value.
In the case where the label value exceeded 255, this meant that the labelmap would be missing from the output.

Fixed by setting the label value for binary labelmap created from closed surface to 1.
This is fine, since the PostConvert step will collapse the labelmaps and assign new label values anyway.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28570 3bd1e089-480b-0410-8dfb-8563597acbee