PERF: Improve Grow from seeds effect speed and decrease memory usage
Simplified design of FibHeap class. Instead of completing heap node implementation in a derived class, implement everything in one class.
This allows removing of virtual methods (making things faster and reduce memory usage) and optimization of basic operators.
Changed how nodes reference other nodes to use unsigned int index (32 bits) instead of a pointer (64 bits), which is significant because
a node is created for each image voxel and there can be hundreds of millions of voxels in an image.
Now the class is less generic, so it is not reusable elsewhere, but its performance is better.

Refactored vtkImageGrowCutSegment to not require storing of previous distances and labels (this information is now deduced from current distances
and nodes).

As a result, peak memory usage of the effect is reduced by about 40% and computation time is reduced by about 20%.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28651 3bd1e089-480b-0410-8dfb-8563597acbee