BUG: Improve segment deep copy to separate shared labelmaps
When copying labelmap representations between segments, all values in the labelmap were previously copied.
This commit changes vtkSegment::DeepCopy to only copy the label value of the source segment.

Reimplements r28672 with some additional changes to ensure that DeepCopy is not used in vtkSegmentationHistory.
This ensures that Undo/Redo behavior is not affected.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28687 3bd1e089-480b-0410-8dfb-8563597acbee