BUG: Fix replace for shared labelmaps with SetBinaryLabelmapToSegment
Calling vtkSlicerSegmentationsModuleLogic::SetBinaryLabelmapToSegment with replace mode previously copied the contents of the modifier labelmap to the segment labelmap without checking to see if it is a shared labelmap.
Fixed by checking if the segment labelmap is shared. If the segment is shared the label value cleared for the segment and masking operations are used instead.

git-svn-id: http://svn.slicer.org/Slicer4/trunk@28752 3bd1e089-480b-0410-8dfb-8563597acbee